OGM-292 Renaming org.hibernate.ogm.datastore.{couchdb|mongodb}.options.mapping to ....options.context
