+++ couchdb/src/main/java/org/hibernate/ogm/datastore/couchdb/CouchDB.java
@@ -20,10 +20,10 @@
  */
 package org.hibernate.ogm.datastore.couchdb;
 
-import org.hibernate.ogm.datastore.couchdb.options.mapping.CouchDBGlobalContext;
-import org.hibernate.ogm.datastore.couchdb.options.mapping.impl.CouchDBEntityContextImpl;
-import org.hibernate.ogm.datastore.couchdb.options.mapping.impl.CouchDBGlobalContextImpl;
-import org.hibernate.ogm.datastore.couchdb.options.mapping.impl.CouchDBPropertyContextImpl;
+import org.hibernate.ogm.datastore.couchdb.options.context.CouchDBGlobalContext;
+import org.hibernate.ogm.datastore.couchdb.options.context.impl.CouchDBEntityContextImpl;
+import org.hibernate.ogm.datastore.couchdb.options.context.impl.CouchDBGlobalContextImpl;
+import org.hibernate.ogm.datastore.couchdb.options.context.impl.CouchDBPropertyContextImpl;
 import org.hibernate.ogm.datastore.spi.DatastoreConfiguration;
 import org.hibernate.ogm.options.navigation.impl.ConfigurationContext;
