+++ couchdb/src/main/java/org/hibernate/ogm/datastore/couchdb/options/context/impl/CouchDBEntityContextImpl.java
@@ -18,14 +18,22 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA  02110-1301, USA.
  */
-package org.hibernate.ogm.datastore.couchdb.options.mapping;
+package org.hibernate.ogm.datastore.couchdb.options.context.impl;
 
-import org.hibernate.ogm.options.navigation.document.DocumentStorePropertyContext;
+import org.hibernate.ogm.datastore.couchdb.options.context.CouchDBEntityContext;
+import org.hibernate.ogm.datastore.couchdb.options.context.CouchDBPropertyContext;
+import org.hibernate.ogm.options.navigation.impl.ConfigurationContext;
+import org.hibernate.ogm.options.navigation.impl.document.DocumentStoreEntityContextImpl;
 
 /**
- * Allows to configure CouchDB-specific options applying on a per-entity level.
+ * Converts CouchDB entity-level options.
  *
  * @author Gunnar Morling
  */
-public interface CouchDBPropertyContext extends DocumentStorePropertyContext<CouchDBEntityContext, CouchDBPropertyContext> {
+public abstract class CouchDBEntityContextImpl extends DocumentStoreEntityContextImpl<CouchDBEntityContext, CouchDBPropertyContext> implements
+		CouchDBEntityContext {
+
+	public CouchDBEntityContextImpl(ConfigurationContext context) {
+		super( context );
+	}
 }
