+++ mongodb/src/main/java/org/hibernate/ogm/datastore/mongodb/options/context/MongoDBGlobalContext.java
@@ -18,7 +18,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA  02110-1301, USA.
  */
-package org.hibernate.ogm.datastore.mongodb.options.mapping;
+package org.hibernate.ogm.datastore.mongodb.options.context;
 
 import org.hibernate.ogm.datastore.mongodb.options.AssociationDocumentType;
 import org.hibernate.ogm.datastore.mongodb.options.WriteConcernType;
