+++ mongodb/src/main/java/org/hibernate/ogm/datastore/mongodb/options/context/MongoDBPropertyContext.java
@@ -18,15 +18,26 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA  02110-1301, USA.
  */
-package org.hibernate.ogm.datastore.couchdb.options.mapping;
+package org.hibernate.ogm.datastore.mongodb.options.context;
 
-import org.hibernate.ogm.options.navigation.document.DocumentStoreGlobalContext;
+import org.hibernate.ogm.datastore.mongodb.options.AssociationDocumentType;
+import org.hibernate.ogm.options.navigation.document.DocumentStorePropertyContext;
 
 /**
- * Allows to configure CouchDB-specific options applying on a global level. These options may be overridden for single
- * entities or properties.
+ * Allows to configure MongoDB-specific options for a single property.
  *
+ * @author Davide D'Alto <davide@hibernate.org>
  * @author Gunnar Morling
  */
-public interface CouchDBGlobalContext extends DocumentStoreGlobalContext<CouchDBGlobalContext, CouchDBEntityContext> {
+public interface MongoDBPropertyContext extends DocumentStorePropertyContext<MongoDBEntityContext, MongoDBPropertyContext> {
+
+	/**
+	 * Specifies how association documents should be persisted. Only applies when the association storage strategy is
+	 * set to {@link org.hibernate.ogm.options.generic.document.AssociationStorageType#ASSOCIATION_DOCUMENT}.
+	 *
+	 * @param associationDocumentStorage the association document type to be used; overrides any settings on the entity
+	 * or global level
+	 * @return this context, allowing for further fluent API invocations
+	 */
+	MongoDBPropertyContext associationDocumentStorage(AssociationDocumentType associationDocumentStorage);
 }
